+++ app/src/components/App.js
@@ -6,22 +6,20 @@ function App() {
   const auth = fbAuth;
 
   const [init, setInit] = useState(false);
-  const [isLoggedIn, setIsLoggedIn] = useState(false);
-  useEffect(() => {    
+  const [userObj, setUserObj] = useState(null);
+  useEffect(() => {
     // reference: https://firebase.google.com/docs/reference/js/v8/firebase.auth.Auth#onauthstatechanged
     auth.onAuthStateChanged((user) => {
       if (user) {
-        setIsLoggedIn(true);        
-      } else {
-        setIsLoggedIn(false);
-      }
+        setUserObj(user);
+      } 
       setInit(true);
     });
   }, [])
 
   return (
     <div className="App">
-      {init ? <AppRouter isLoggedIn={isLoggedIn} /> : "Loading..." }
+      {init ? <AppRouter isLoggedIn={Boolean(userObj)} userObj={userObj} /> : "Loading..."}
     </div>
   );
 }